updated webpages

+++ greetingsapp/OutputWithDb.cs
@@ -20,23 +20,23 @@ public class OutputWithDb : IOutput
         }
     }
     //"Data Source=./trying_dapper.db";
-    public  IEnumerable<UserModel> GetUsers()
+    public IEnumerable<UserModel> GetUsers()
     {
         using (var connection = new SqliteConnection(_connectionString))
         {
-           var users = connection.Query<UserModel>(@"select * from user");
-           return users;
+            var users = connection.Query<UserModel>(@"select * from user");
+            return users;
         }
     }
-    public  UserModel GetUser(string Name)
+    public UserModel GetUser(string Name)
     {
-        var template = new UserModel {Name = Name};
+        var template = new UserModel { Name = Name };
         var parameters = new DynamicParameters(template);
         var sql = @"select * from user where Name = @Name";
         using (var connection = new SqliteConnection(_connectionString))
         {
-           var user = connection.QueryFirstOrDefault<UserModel>(sql,parameters);
-           return user;
+            var user = connection.QueryFirstOrDefault<UserModel>(sql, parameters);
+            return user;
         }
     }
 
@@ -46,35 +46,52 @@ public class OutputWithDb : IOutput
         {
             var sql = @" insert into  user (Name, Count, Language)
 	                    values (@Name, @Count, @Language);";
-            var parameters =  new UserModel() { Name = user.Name, Count = user.Count,
-	            Language = user.Language};
+            var parameters = new UserModel()
+            {
+                Name = user.Name,
+                Count = user.Count,
+                Language = user.Language
+            };
 
-           connection.Execute( sql, parameters);
+            connection.Execute(sql, parameters);
         }
     }
 
-    public void Update(UserModel  user)
+    public void Update(UserModel user)
     {
         using (var connection = new SqliteConnection(_connectionString))
         {
-            var template = new UserModel { Name = user.Name, Count = user.Count+1, Language = user.Language};
+            var template = new UserModel { Name = user.Name, Count = user.Count + 1, Language = user.Language };
             var parameters = new DynamicParameters(template);
             var sql = @" update user
 	                     set Count= @Count, Language = @Language
                          where Name = @Name;";
-           connection.Execute( sql, parameters);
-        }   
+            connection.Execute(sql, parameters);
+        }
     }
-    public void Remove(string Name)
+    public void Remove(UserModel user)
     {
         using (var connection = new SqliteConnection(_connectionString))
         {
-            var template = new UserModel { Name = Name};
-            var parameters = new DynamicParameters(template);
-            var sql = @"delete from user
+            if (user.Count > 1)
+            {
+                var template = new UserModel { Name = user.Name, Count = user.Count - 1 };
+                var parameters = new DynamicParameters(template);
+                var sql = @"update user
+                          set Count= @Count
+                          where Name = @Name;";
+                connection.Execute(sql, parameters);
+            }
+            else
+            {
+                var template = new UserModel { Name = user.Name };
+                var parameters = new DynamicParameters(template);
+                var sql = @"delete from user
                         where Name = @Name;";
-           connection.Execute( sql, parameters);
-        }   
+                connection.Execute(sql, parameters);
+            }
+
+        }
     }
 
     public void Clear()
@@ -82,8 +99,8 @@ public class OutputWithDb : IOutput
         using (var connection = new SqliteConnection(_connectionString))
         {
             var sql = @"delete from user;";
-           connection.Execute(sql);
-        }   
+            connection.Execute(sql);
+        }
     }
 
     public void Help()
